Playlist functionality, Google Sheets integration, and more
Added the option to automatically add videos to a plalist on YouTube as they are created.
Playlists can be selected from a list or created from the settings tab.
Functionality to use Google Sheets added- names, teams, and twitter tags are imported from a sheet specified in the settings.
Youtube settings and Misc Settings are now seperate tabs.
Thumbnails are now saved as seperate filed to allow for side-by-side uploading.
Google Sheets needs to have functionality added to choose a character, which also needs a drop down box for color on the main page.