Settings Update
Full functionality seemingly in place for google sheets reading and writing for player information
Work on updating and expanding the settings pab into sub-tabs started.
This will allow for new settings to be added as needed into their own respective subtabs, such as thumbnail graphical settings.
left subtabs seem to be finished
right subtabs need to be fleshed out, but appearance is complete