TournamentDataLibrary created Created a new library and moved sql-related content here. Meant to hold all data related to the tournament.
Started this since the information used to generate a thumbnail would be almost all of a match's information. It'd be better to establish a data type that can be used across the entire app.

On top of this, other pieces of tournament-related data needs refactored to fit into this model. Having this clear will be able to make the methods goals of parts of the app more clear since every part relies on tournament data.